+++ M01/Assignment/app/src/main/java/com/example/books/DetailActivity.kt
@@ -1,22 +1,20 @@
 package com.example.books
 
-import android.app.Activity
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.text.TextWatcher
 import kotlinx.android.synthetic.main.activity_detail.*
 
 class DetailActivity : AppCompatActivity() {
 
-    private var entry: Book = Book(Book.INVALID_ID)
+    private var entry: BookEntry = BookEntry(BookEntry.INVALID_ID)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_detail)
 
         val intent = intent
-        entry = intent.getSerializableExtra(Book.TAG) as Book
+        entry = intent.getSerializableExtra(BookEntry.TAG) as BookEntry
 
         entry_id_label.text = "${entry.id}"
         journal_entry_text.setText(entry.title)
