+++ M01/Assignment/app/src/main/java/com/example/books/MainActivity.kt
@@ -5,6 +5,8 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import com.example.books.Book.Companion.createBookEntry
+import com.example.books.Book.Companion.createBookEntry1
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
@@ -17,7 +19,7 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    private var entryList = mutableListOf<Book>()
+    private var entryList = mutableListOf<BookEntry>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,11 +27,11 @@ class MainActivity : AppCompatActivity() {
 
         button.setOnClickListener {
             val intent = Intent(this@MainActivity, DetailActivity::class.java)
-            val entry = BookEntry() as Book
-            intent.putExtra(Book.TAG, entry)
+            val entry = createBookEntry()
+            intent.putExtra(BookEntry.TAG, entry)
             startActivityForResult(intent, NEW_ENTRY_REQUEST)
         }
-        entryList = prefs?.readALL()
+        entryList = prefs.readALL()
     }
 
     override fun onStart() {
@@ -45,7 +47,7 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    private fun createBookEntry(entry: Book): TextView {
+    private fun createBookEntry(entry: BookEntry): TextView {
         val view = TextView(this@MainActivity)
         view.text = "${entry.id}, ${entry.reasonToRead}, ${entry.title}, ${entry.hasBeenRead}"
         view.setPadding(15, 15, 15, 15)
@@ -53,7 +55,7 @@ class MainActivity : AppCompatActivity() {
 
         view.setOnClickListener {
             val viewDetailIntent = Intent(this@MainActivity, DetailActivity::class.java)
-            viewDetailIntent.putExtra(Book.TAG, entry)
+            viewDetailIntent.putExtra(BookEntry.TAG, entry)
             startActivityForResult(viewDetailIntent, EDIT_ENTRY_REQUEST)
         }
         return view
@@ -65,13 +67,13 @@ class MainActivity : AppCompatActivity() {
         if (requestCode== Activity.RESULT_OK){
             if (requestCode== NEW_ENTRY_REQUEST){
                 if (data != null){
-                    val entry = data.getSerializableExtra(Book.TAG) as Book
+                    val entry = data.getSerializableExtra(BookEntry.TAG) as BookEntry
                     entryList.add(entry)
                     prefs.updateEntry(entry)
                 }
             }else if(requestCode== EDIT_ENTRY_REQUEST){
                 if (data != null){
-                    val entry = data.getSerializableExtra(Book.TAG) as Book
+                    val entry = data.getSerializableExtra(BookEntry.TAG) as BookEntry
                     entryList[entry.id] = entry
                     prefs.updateEntry(entry)
                 }
