+++ M01/Assignment/app/src/main/java/com/example/books/Prefs.kt
@@ -6,7 +6,7 @@ import java.lang.StringBuilder
 
 class Prefs(context: Context){
     companion object{
-        private const val BOOK_PREFS = "Book prefs"
+        private const val BOOK_PREFS = "BookEntry prefs"
         private const val NEXT_ID_KEY = "Next ID"
         private const val ID_LIST_KEY = "Id list key"
         private const val ENTRY_PREFIX = " entry prefix_"
@@ -14,10 +14,10 @@ class Prefs(context: Context){
     // val for sharedPrefs
     val sharedPrefs: SharedPreferences = context.getSharedPreferences(BOOK_PREFS, Context.MODE_PRIVATE)
 
-    fun createBookEntry(entry: Book){
+    fun createBookEntry(entry: BookEntry){
         //read list of ids
         val ids = getListofIDS()
-        if (entry.id == Book.INVALID_ID && !ids.contains(entry.id.toString())){
+        if (entry.id == BookEntry.INVALID_ID && !ids.contains(entry.id.toString())){
             //new entry
             val editor = sharedPrefs.edit()
                 var nextID = sharedPrefs.getInt(NEXT_ID_KEY, 0)
@@ -50,27 +50,29 @@ class Prefs(context: Context){
         return ids
     }
     //read an existing entry
-    fun readExisting(id: Int): Book{
+    fun readExisting(id: Int): BookEntry{
         val entryCsv = sharedPrefs.getString(ENTRY_PREFIX + id, "invalid") ?: " "
         return entryCsv?.let {
-            Book(entryCsv)
+            BookEntry(entryCsv)
         }
     }
     //read all entries
-    fun readALL():MutableList<Book>{
+    fun readALL():MutableList<BookEntry>{
         //read the list of ids
         val listofIDS = getListofIDS()
         //step through and read each entry
-        val entryList =java.util.ArrayList<Book>()
-        for (id in listofIDS){
-            readExisting(id.toInt())?.let {
-                entryList.add(it)
+        val entryList =java.util.ArrayList<BookEntry>()
+        for (id in listofIDS) {
+            if (id.isNotBlank()) {
+                readExisting(id.toInt())?.let {
+                    entryList.add(it)
+                }
             }
-            }
+        }
             return entryList
         }
 
-    fun updateEntry(entry: Book){
+    fun updateEntry(entry: BookEntry){
         val editor = sharedPrefs.edit()
         editor.putString(ENTRY_PREFIX + entry.id, entry.toCsvString())
         editor.apply()
